services/admin_panel/pyproject.toml
[tool.poetry]
name = "notifications_sprint_1"
version = "0.1.0"
description = ""
authors = ["GA10"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "1.10.2"
django = "^4.1.7"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
django-split-settings = "^1.2.0"
psycopg2-binary = "^2.9.5"
uwsgi = {version = "^2.0.21", platform = "linux"}
django-tinymce = "^3.5.0"
fastapi = "^0.92.0"
pika = "^1.3.1"
aio-pika = "^9.0.4"
uvicorn = "^0.20.0"
orjson = "^3.8.7"
asyncio = "^3.4.3"
jinja2 = "^3.1.2"
faker = "^17.6.0"
pyjwt = "^2.6.0"
requests = "^2.28.2"
aiohttp = "^3.8.4"
sentry-sdk = "^1.16.0"
python-logstash = "^0.4.8"
sqlalchemy = "1.4.45"
databases = "^0.7.0"
aioredis = "^2.0.1"
python-jose = "^3.3.0"
types-requests = "^2.28.11.15"


[tool.poetry.dev-dependencies]
flake8-bugbear = "^22.6.22"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.3.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-variables-names = "^0.0.5"
pre-commit = "^2.19.0"
dlint = "^0.13.0"
black = "^22.6.0"
isort = "^5.10.1"
yesqa = "^1.4.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
pytest-asyncio = "^0.19.0"
aiohttp = "^3.8.1"
flake8-html = "^0.4.3"
celery-stubs = "^0.1.3"


[tool.black]
line-length = 120
target-version = ["py38"]
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
      \.venv
    | \.git
    | \.env
  )/
)
'''

[tool.isort]
multi_line_output = 3
line_length = 120
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
